tests/requirements.txt
selenium
webdriver_manager
pillow
pyyaml

